+++ client/src/components/Admin/Home/home.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function AdminHome () {
+  return (
+    <h1>Admin panel</h1>
+  )
+}
+export default AdminHome;
